requirements.txt
alembic==1.7.5
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
defusedxml==0.7.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
httpie==2.6.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
#psycopg2==2.9.3
Pygments==2.11.2
PyJWT==2.3.0
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.8
Werkzeug==2.0.2
